per conversation with Karl, remove giftopng.
it is still referenced from Master/texmf-dist/tex/xmltex/passivetex/fotex.sty:

  \@namedef{Gin@rule@.gif}#1{{png}{.png}{`giftopng #1}}

but the only people using it would be

(a) using xmltex
(b) and using gif files
(c) and not having corresponding png's (or something) already,
(d) and having \write18 enabled

My guess is that the set of such people in the world is empty.
My thought is, let's leave it and see if anyone notices :).