Ignore empty file in filetype checking
On OS X 10.6 at least opening non-exist resource fork of a file with /rsrc
suffix by fopen() will not return NULL but instead you will get a valid
FILE* but you can't read anything. It regressed in r36978, probably
because we didn't try to check for dfonts like that before the change.