rework (again) location of various configuration options
Currently opt_frozen is set in 00texlive.installation and thus
was in principle targeted for user editing. Furthermore, it wasn't
a defined key in TLConfig::TLPDBOptions.

The frozen setting has now moved in 00texlive.config.tlpsrc, which
also defines the container format etc, thus the proper place.

At the same time we cleared out 00texlive.installation.tlpsrc,
it is not necessary anymore, and only carried wrong information.
All the defaults for options/settings in 00texlive.installation
are now defined in TLConfig.